refactor(money): promote parallax value mapping to Money utils (MUSD-1204)
The tilt-to-Rive mapping is no longer specific to the next-best-action card: the First Deposit screen drives the same xValue/yValue bindings. Move it beside the other Money utils so both surfaces import it from a shared location rather than reaching into a sibling component's folder.
