+++ app/components/UI/Money/utils/parallax.ts
